requirements.txt
streamlit==1.32.0
tensorflow==2.16.1
numpy==1.26.4
Pillow==10.3.0
cryptography==42.0.5